requirements.txt
pyspark
kafka-python
pyarrow
pandas
apache-flink
